% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_scatterplot.R
\name{ADVANCDED_SCATTER_SERVER}
\alias{ADVANCDED_SCATTER_SERVER}
\title{ADVANCDED_SCATTER_SERVER}
\usage{
ADVANCDED_SCATTER_SERVER(
  id,
  data_frame,
  x_index = 1,
  y_index = 2,
  grouping_index = NULL
)
}
\arguments{
\item{grouping_index}{}
}
\value{

}
\description{
ADVANCDED_SCATTER_SERVER
}
\examples{

}
