% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_scatterplot.R
\name{ADVANCDED_SCATTER_UI}
\alias{ADVANCDED_SCATTER_UI}
\title{ADVANCDED_SCATTER_UI}
\usage{
ADVANCDED_SCATTER_UI(
  id,
  group_coloring = FALSE,
  group_shaping = FALSE,
  collapsed = FALSE
)
}
\arguments{
\item{collapsed}{}
}
\value{

}
\description{
ADVANCDED_SCATTER_UI
}
\examples{

}
