% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_display_datatable.R
\name{DISPLAY_TABLE_SERVER}
\alias{DISPLAY_TABLE_SERVER}
\title{shiny module for datatable display, server end}
\usage{
DISPLAY_TABLE_SERVER(
  input,
  output,
  session,
  data_table,
  filename_tag,
  height = 600
)
}
\arguments{
\item{height}{}
}
\value{

}
\description{
shiny module for datatable display, server end
}
