% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_display_datatable.R
\name{DISPLAY_TABLE_SIMPLE_SERVER}
\alias{DISPLAY_TABLE_SIMPLE_SERVER}
\title{shiny module for table display, whithout download button, server end}
\usage{
DISPLAY_TABLE_SIMPLE_SERVER(input, output, session, data_table, height = 600)
}
\arguments{
\item{height}{}
}
\value{

}
\description{
shiny module for table display, whithout download button, server end
}
