% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_display_datatable.R
\name{DISPLAY_TABLE_SIMPLE_UI}
\alias{DISPLAY_TABLE_SIMPLE_UI}
\title{shiny module for table display, without download button, ui end}
\usage{
DISPLAY_TABLE_SIMPLE_UI(id)
}
\arguments{
\item{id}{}
}
\value{

}
\description{
shiny module for table display, without download button, ui end
}
