% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_display_datatable.R
\name{DISPLAY_TABLE_UI}
\alias{DISPLAY_TABLE_UI}
\title{shiny module for datatable display, ui end}
\usage{
DISPLAY_TABLE_UI(id, boxtitle = NULL, boxwidth = 12)
}
\arguments{
\item{boxwidth}{}
}
\value{

}
\description{
ui has download button
}
\examples{

}
