% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_download_table.R
\name{DOWNLOAD_TABLE_SERVER}
\alias{DOWNLOAD_TABLE_SERVER}
\title{shiny module to download table, server end}
\usage{
DOWNLOAD_TABLE_SERVER(input, output, session, data, filename_tag = NULL)
}
\arguments{
\item{filename_tag}{}
}
\value{

}
\description{
shiny module to download table, server end
}
