% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_download_table.R
\name{DOWNLOAD_TABLE_UI}
\alias{DOWNLOAD_TABLE_UI}
\title{shiny module to download table, ui end}
\usage{
DOWNLOAD_TABLE_UI(id, label = "Download CSV")
}
\arguments{
\item{label}{}
}
\value{

}
\description{
shiny module to download table, ui end
}
