% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor2color.R
\name{factor2color}
\alias{factor2color}
\title{low level function convert a vector to corresponding colors, useful for marking heatmap grouping
available colorschemes from Rcolor Brewer qulity colors.
see RColorBrewer::brewer.pal.info for detailed options and colors}
\usage{
factor2color(
  factor_input,
  RcolorBrewer_theme = "Spectral",
  Rcolorbrewer_schemes = "div"
)
}
\arguments{
\item{factor_input}{a vector of factor}

\item{RcolorBrewer_theme}{options from "Accent"  "Dark2"   "Paired"  "Pastel1" "Pastel2" "Set1"    "Set2"    "Set3"}

\item{Rcolorbrewer_schemes}{one of c("seq","qual","div")}
}
\value{
a vector of colors, the the same levels of the input vector
}
\description{
low level function convert a vector to corresponding colors, useful for marking heatmap grouping
available colorschemes from Rcolor Brewer qulity colors.
see RColorBrewer::brewer.pal.info for detailed options and colors
}
\examples{
f <- as.factor(sample(LETTERS[1:3], 10, replace = TRUE))
factor2color(f, "Spectral")



}
