% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor2colorDF.R
\name{factor2colorDF}
\alias{factor2colorDF}
\title{low level function convert a dataframe of factors to corresponding colors, useful for marking heatmap/HCA grouping
avaliable colorschemes from Rcolor Brewer qulity colors.}
\usage{
factor2colorDF(
  DF,
  RcolorBrewer_theme = "Spectral",
  Rcolorbrewer_schemes = "div"
)
}
\arguments{
\item{DF}{a data.frame of factors (or elements which can be converted to factor)}

\item{RcolorBrewer_theme}{options from "Accent"  "Dark2"   "Paired"  "Pastel1" "Pastel2" "Set1"    "Set2"    "Set3"}

\item{Rcolorbrewer_schemes}{one of c("seq","qual","div")}
}
\value{
a data.frame of color codes
}
\description{
low level function convert a dataframe of factors to corresponding colors, useful for marking heatmap/HCA grouping
avaliable colorschemes from Rcolor Brewer qulity colors.
}
\examples{
factor1 <- sample(LETTERS[1:3], 10, replace = TRUE)
factor2 <- sample(LETTERS[1:3], 10, replace = TRUE)
factor2colorDF(cbind(factor1, factor2))


}
