% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_experiment_names.R
\name{find_experiment_names}
\alias{find_experiment_names}
\title{Extract experiment names from Metalab result table}
\usage{
find_experiment_names(data_table)
}
\arguments{
\item{data_table}{could be peptide, protein, taxon and function table}
}
\value{
a list, with both index and names, and the data.table with new experiment names
}
\description{
In Metablab, the quantity information (intensity) are organized in result tables, with column headers usually starting with names like LFQ intensity (from LFQ quantification), Intensity (from raw data, labeled data, including TMT) etc. This function can automatically extract the experiment names for easy plotting or checking if table are consistent with meta_table provided.
}
\examples{

}
