% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_lipid_combo.R
\name{find_lipid_combo}
\alias{find_lipid_combo}
\title{find 3 chains combo for a lipid structure}
\usage{
find_lipid_combo(combo_target, chains_observed)
}
\arguments{
\item{combo_target}{a string, general format as carbon number:double bound number, like 51:4}

\item{chains_observed}{a string vector, consists of all observed fatty acid chains, with  general format as carbon number:double bound number,}
}
\value{
a data.frame, with all possible 3 chain compositions. The order of the chain does not matter.
}
\description{
find_lipid_combo start with more than 3 observed fatty acid chains with defined carbo and double bound number to find the
combo to achieve a target lipid structure. The three chains could be the same (replacement sampling).
}
\examples{

}
\seealso{
\code{\link{combn}}
}
