% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_p_location.R
\name{find_p_location}
\alias{find_p_location}
\title{locate the significant pairs from ANOVA test result
internal function}
\usage{
find_p_location(p_matrix, p_threshold = 0.05)
}
\arguments{
\item{p_matrix}{p value matrix from pairwise.t.test like functions (need to $p.value)}

\item{p_threshold}{pvalue threshold}
}
\value{
list of the group pairs
}
\description{
locate the significant pairs from ANOVA test result
internal function
}
\examples{
attach(airquality)
Month <- factor(Month, labels = month.abb[5:9])
result<-pairwise.t.test(Ozone, Month, p.adj = "fdr")
find_p_location(result$p.value)




}
