% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_colors.R
\name{generate_colors}
\alias{generate_colors}
\title{Wrapper of RColorBrewer to produce any number of colors by colorRampPalette}
\usage{
generate_colors(
  ncolors = 10,
  Rcolorbrewer_schemes = "div",
  RcolorBrewer_theme = "Spectral"
)
}
\arguments{
\item{ncolors}{nubmer of colors to generate}

\item{Rcolorbrewer_schemes}{one of c("seq","qual","div") div means diverging, qual means qualititive, both with high contract, while seq means sequential, from the same color tone, good for quantative}

\item{RcolorBrewer_theme}{color theme, check all names by brewer.pal.info or displayed by display.brewer.all()}
}
\value{
a vector of color codes
}
\description{
For scheme "seq",  check the available options by: rownames(brewer.pal.info[brewer.pal.info$category == "seq",])
For scheme "qual", check by: rownames(brewer.pal.info[brewer.pal.info$category == "qual",])
For scheme "div", check by: rownames(brewer.pal.info[brewer.pal.info$category == "div",])
}
\examples{
generate_colors(100)
col <- generate_colors(100, "seq","Spectral" )
pie(rep(1,100), col = col,border = FALSE,labels = "")

}
