% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_prettier.R
\name{ggplot2_prettier}
\alias{ggplot2_prettier}
\title{ggplot2_prettier}
\usage{
ggplot2_prettier(
  ggplot2_object,
  maintitle = NULL,
  xlab = NULL,
  ylab = NULL,
  axis.text.angle.x = 0,
  axis.text.angle.y = 0,
  vertical = FALSE
)
}
\arguments{
\item{vertical}{}
}
\value{

}
\description{
ggplot2_prettier
}
