% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_ID_SUMMARY.R
\name{ID_SUMMARY_SERVER}
\alias{ID_SUMMARY_SERVER}
\title{shiny module for ID summary, server end}
\usage{
ID_SUMMARY_SERVER(input, output, session, data_table, meta_table = NULL)
}
\arguments{
\item{data_table}{a clean data_frame with rownames and columnames}

\item{meta_table}{}
}
\value{

}
\description{
shiny module for ID summary, server end
}
