% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_ID_SUMMARY.R
\name{ID_SUMMARY_UI}
\alias{ID_SUMMARY_UI}
\title{shiny module for ID summary, ui end}
\usage{
ID_SUMMARY_UI(id)
}
\arguments{
\item{id}{}
}
\value{



}
\description{
shiny module for ID summary, ui end
}
\examples{
shiny module for ID summary, ui end

}
