% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.packages.auto.r
\name{install.packages.auto}
\alias{install.packages.auto}
\title{instal packages automatically}
\usage{
install.packages.auto(x)
}
\arguments{
\item{x}{package name, with or without the quote both works}
}
\value{
None The same as intall.packages
}
\description{
This function will automatically install and library/load the package when invoked.
It is a quick wrapper of of the install.package and biocoductor bioLite.
It also does not need quote on the package name.
}
\details{
This function is just for quick and easy install and library usage. It does not consider installation from github yet.
If the target pacakge has already been installed, no installation will be proceeded.
With R version 3.5 or greater, Bioconductor packages installing start using BiocManager. Therefore this comand only woorks for biocondcutior for R 3.5 or greater
}
\examples{
install.packages.auto(roxygen2) # from CRAN
install.packages.auto(qvalue) # from bioconductor
}
\seealso{
\code{\link{install.packages}}
}
