% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isValidEmail.R
\name{isValidEmail}
\alias{isValidEmail}
\title{check if a string as a valid email address}
\usage{
isValidEmail(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
check if a string as a valid email address
}
