% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_Anova_1_PostHoc.R
\name{matrix_Anova_1_PostHoc}
\alias{matrix_Anova_1_PostHoc}
\title{wrapper to apply Anova and post hoc on data matrix in a rowwise way}
\usage{
matrix_Anova_1_PostHoc(data, groups)
}
\arguments{
\item{data}{data matrix}

\item{groups}{factor of grouping}
}
\value{
a data frame
}
\description{
wrapper to apply Anova and post hoc on data matrix in a rowwise way
}
