% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQ_QC_plot.R
\name{MQ_QC_plot}
\alias{MQ_QC_plot}
\title{a ggplot2 wrapper plot MS identification rate check}
\usage{
MQ_QC_plot(
  data.frame,
  plot_type = c("scatter", "bar", "density", "histogram", "freqpoly", "box", "violin"),
  group = NULL,
  cutoff = 20,
  maintitle = "",
  xlabel = "",
  vertical = FALSE,
  ...
)
}
\arguments{
\item{data.frame}{a data from with first two columns to be used, first as name, second as MS id rate}

\item{plot_type}{options are one or more in a c("scatter", "bar", "density", "histogram", "freqpoly", "box", "violin"), defautl as bar}

\item{group}{a vector of grouping information, the same order as the rows}

\item{cutoff}{a cutoff indicator for threshold plot, indicating usually accepted bar}

\item{maintitle}{plot title}

\item{xlabel}{label for x}

\item{vertical}{if too many lines, vertical layout holds more information.}

\item{...}{}
}
\value{
list of selected plot
}
\description{
a ggplot2 wrapper plot MS identification rate check
}
\examples{

# see example for input data format
my_test <- data.frame("samplename" = paste0("sample_", 1:20),
                      "msms_id" = c(abs(rnorm(10))*10+20, abs(rnorm(10))*10+30),
                      "treat_group" = c(paste0("group_", rep("A",10)), paste0("group_", rep("B",10)))
)

tt <-  MQ_QC_plot(my_test, plot_type = c("scatter","bar","density", "histogram", "freqpoly", "box", "violin"), cutoff = 35, group = "treat_group", maintitle = "MSMS ID Rate", xlabel = "MS/MS ID \%")
# now tt has all the required plot
tt$scatter
tt$density




}
