% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_plot_with_confidence_2.R
\name{PCA_plot_with_confidence_2}
\alias{PCA_plot_with_confidence_2}
\title{Title}
\usage{
PCA_plot_with_confidence_2(prcomp_out, data_meta)
}
\arguments{
\item{data_meta}{}
}
\value{

}
\description{
Title
}
