% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_beta_diversity.R
\name{plot_beta_diversity}
\alias{plot_beta_diversity}
\title{plot beta divesity using PCoA}
\usage{
plot_beta_diversity(count_species, meta_table = NULL)
}
\arguments{
\item{meta_table}{}
}
\value{

}
\description{
plot beta divesity using PCoA
}
