% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_box_matrix_byrow.R
\name{plot_box_matrix_byrow}
\alias{plot_box_matrix_byrow}
\title{boxPlot a row in a matrix}
\usage{
plot_box_matrix_byrow(
  row = 1,
  data_matrix,
  data_meta = NULL,
  plot_type = "violin"
)
}
\arguments{
\item{row}{index of the row to plot}

\item{data_matrix}{input of the data matrix}

\item{data_meta}{optional, a data.frame of meta, first column as sample name corresponding to the column name of the data matrix, second column as grouping inforation}

\item{plot_type}{string, either  violin or boxplot}
}
\value{
a ggplot2 object
}
\description{
boxPlot a row in a matrix
}
\examples{

# prepare data
t <- generate_test_data()

# plot a single row
plot_box_matrix_byrow(row = 1, data_matrix = t$matrix, data_meta = t$meta, plot_type = "violin")

# plot multiple rows into a list
plot_multiple_rows <- lapply(1:10, plot_box_matrix_byrow, data = t$matrix, data_meta = t$meta)


}
