% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_box_vector.R
\name{plot_box_vector}
\alias{plot_box_vector}
\title{Boxplot from a vector along with grouping}
\usage{
plot_box_vector(
  vector = NULL,
  factor = 1,
  plot_type = "violin",
  title = "Distribution"
)
}
\arguments{
\item{vector}{a numeric vector}

\item{factor}{a factor vector, the same order as the vector}

\item{plot_type}{type of plot, default as "violin", alternative as "boxboxplot"}

\item{title}{a string of title name}
}
\value{
a ggplot2 object of either boxplot or violin plot
}
\description{
A wrapper function of geom_violin or geom_boxplo. Will do the data data,framn preaparation first
depricated name: simple_ggboxplot_vector
}
\examples{

vector_test <- rnorm(200)
factor_test <- sample(LETTERS[1:3], 200, replace = TRUE)
plot_box_vector(vector_test, factor_test)
plot_box_vector(vector = vector_test, factor = factor_test, plot_type = "boxplot")




}
