% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_HCA.R
\name{plot_HCA}
\alias{plot_HCA}
\title{plot HCA from intensity matrix}
\usage{
plot_HCA(df_intensity = NULL, meta_table = NULL, plus1 = TRUE)
}
\arguments{
\item{df_intensity}{data matrix with MS raw readings}

\item{meta_table}{data.frame for grouping}

\item{plus1}{TRUE or FALSE, means if all values +1 for easy log transformation for intensitymatrix}
}
\value{
a HCA plot
}
\description{
intensity matrix needs to be along table, with no extra columns, values are MS intensities
the default workflow will add 1 to all density values to make full usage of the whole table, then do log10 transformation, then scale the features, cacluate the distance and plot the dengrogam. will add a colorbar if meta grouping provided
}
