% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_plot_In_box.R
\name{PLOT_IN_BOX_SERVER}
\alias{PLOT_IN_BOX_SERVER}
\title{shiny module to plot in a box, server end}
\usage{
PLOT_IN_BOX_SERVER(
  input,
  output,
  session,
  plot_object,
  plot_type,
  name_tag,
  interactive = TRUE
)
}
\arguments{
\item{name_tag}{for the file name of downloading}

\item{interactive}{}
}
\value{

}
\description{
shiny module to plot in a box, server end
}
