% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_plot_In_box.R
\name{PLOT_IN_BOX_UI}
\alias{PLOT_IN_BOX_UI}
\title{shiny module to plot in a box, ui end}
\usage{
PLOT_IN_BOX_UI(id, boxtitle = NULL, boxwidth = 6, plot_height = 800)
}
\arguments{
\item{boxwidth}{}
}
\value{

}
\description{
shiny module to plot in a box, ui end
}
