% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_KEGG_enrichemt.R
\name{plot_KEGG_enrichemt}
\alias{plot_KEGG_enrichemt}
\title{Plot classical KEGG enrichment graph using basic r command}
\usage{
plot_KEGG_enrichemt(df.match, title.tag = NULL, topN = 20, margin_left = 10)
}
\arguments{
\item{df.match}{data.fram imput, with 4 columns "FunctionName"    "Number_in_group" "Number_matched"  "pvalue"          "pvalue_adjusted"}

\item{title.tag}{A pre in the figure title}

\item{topN}{the number of top N of pvalue_adjusted to plot}

\item{margin_left}{if you have a long FunctionName, you might need to set the margin_left big enought to}
}
\value{
a plot
}
\description{
Plot classical KEGG enrichment graph using basic r command
}
\examples{

# prepare a fake data
size <- sample(100:1000, 10)
p = exp(-abs(sample(1:20,10)))
data <- data.frame(FunctionName =  paste0("KEGG_",1:20 ),
                  Number_in_group = sample(100:1000, 10),
                  Number_matched = size * (sample(1:100, 10)/100),
                 pvalue = p,
                 pvalue_adjusted = p*(sample(1:100, 10)))
# plot
plot_KEGG_enrichemt(data, topN = 10)



}
