% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prcomp_opencpu_csv.r
\name{prcomp_opencpu_csv}
\alias{prcomp_opencpu_csv}
\title{pca test for opencpu development}
\usage{
prcomp_opencpu_csv(filename)
}
\arguments{
\item{combo_target}{a csv file with row as observations, while column as experiment}
}
\value{
a list of sample name, and x,y coordinate
}
\description{
a wrapper for pca test, starting from a csv imput, do the pca analysis, output some txt test.
the coordinate of the pca plot will be returned
}
\seealso{
\code{\link{prcomp}}
}
