% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickheatmap.R
\name{quickheatmap}
\alias{quickheatmap}
\title{simple wrapper of heatmap.2, for quick and easy heatmap plot from data matrix and meta info, with only limited parameters exposed}
\usage{
quickheatmap(
  matrix,
  column_cluster = TRUE,
  column_meta = NULL,
  row_meta = NULL,
  scale = "row",
  color = bluered
)
}
\arguments{
\item{matrix}{a data.frame/matrix}

\item{column_cluster}{Boolean, Switch to do column clustering of not}

\item{column_meta}{a data.frame with 2 columns, 1st as column variables, 2nd as grouping, will map to colors}

\item{row_meta}{a data.frame with 2 columns, 1st as row variables, 2nd as grouping, will map to colors}

\item{scale}{The default is "none", scaling method, options of c("none","row", "column")}

\item{color}{one of redblue,bluered, greenred, redgreen}
}
\value{

}
\description{
20220216 reformat it into rmetalab
20191002 add a line to replace the infinite values, heatmap.2 can deal with NA, but not infinte values
}
\examples{
test_data <- generate_test_data()
quickheatmap(test_data$matrix)
quickheatmap(test_data$matrix, column_cluster = FALSE)
quickheatmap(test_data$matrix, column_meta = test_data$meta[,1:2])
quickheatmap(test_data$matrix, column_meta = test_data$meta[,1:2], column_cluster = FALSE) # used to check the pattern without column clustering

}
