% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickpheatmap.R
\name{quickpheatmap}
\alias{quickpheatmap}
\title{Simple and quick wrapper of ComplexHeatmap::pheatmap,}
\usage{
quickpheatmap(
  matrix,
  column_cluster = TRUE,
  column_meta = NULL,
  row_meta = NULL,
  scale = "row",
  show_rownames = FALSE,
  interactive = FALSE,
  color = colorRampPalette(rev(brewer.pal(n = 7, name = "RdYlBu")))(100)
)
}
\arguments{
\item{matrix}{a data.frame/matrix}

\item{column_cluster}{Boolean, Switch to do column clustering of not}

\item{column_meta}{a data.frame, 1st as column names of the matrix, from 2nd as grouping,supports multiple grouping mark, will map to colors automatically, Null or NA means no mapping}

\item{row_meta}{a data.frame, 1st as row names of the matrix, from 2nd as grouping, supports multiple grouping mark, will map to colors automatically,Null or NA means no mapping}

\item{scale}{The default is "none", scaling method, options of c("none","row", "column")}

\item{color}{a vector of colors, could be generated by any method}
}
\value{
a plot of complex heatmap
}
\description{
For quick and easy heatmap plot from data matrix and meta info, with only limited parameters exposed.
Take advantage of the multiple annotation, easier to implement. The pheatmap:pheatmap is not compatible with markdown document and shiny apps (the ouput postion of the plot is random), thereofore using ComplexHeatmap::pheatmap instead, which is a direct translation of the pheatmap:pheatmap
heatmaply::heatmaply is used  to produce an interactive d3 heatmap when
}
\examples{
test_data <- generate_test_data()
quickpheatmap(test_data$matrix)
quickpheatmap(test_data$matrix, interactive = TRUE)
quickpheatmap(test_data$matrix, color = rainbow(100))
quickpheatmap(test_data$matrix,column_cluster = FALSE)
quickpheatmap(test_data$matrix,column_cluster = FALSE,scale = "column")
quickpheatmap(test_data$matrix,column_cluster = FALSE,scale = "column",column_meta = test_data$meta )
quickpheatmap(test_data$matrix,column_cluster = TRUE,scale = "row",column_meta = test_data$meta )
quickpheatmap(test_data$matrix,column_cluster = TRUE,scale = "row",column_meta = test_data$meta,interactive = TRUE )

}
