% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_plot_upsetR.R
\name{report_plot_upsetR}
\alias{report_plot_upsetR}
\title{generate a plot list with upsetR, to compare with columns, groups if meta provided}
\usage{
report_plot_upsetR(df_intensity, meta_table = NULL)
}
\arguments{
\item{df_intensity}{a data.frame of intensity matrix, with row names defined}

\item{meta_table}{a meta data with right order of grouping, no order check inside the function}
}
\value{
a plot list
}
\description{
This function is for rmarkdonwn report purpose. If meta/grouping information provided, it will generate all comparisons for all groupings, otherwise the plot for all across the sample/columns
}
