% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot_ggrepel.R
\name{scatterplot_ggrepel}
\alias{scatterplot_ggrepel}
\title{scatterplot_ggrepel, a wrapper of scatter plot with ggplot2, and repel function, easy to map color shape etc}
\usage{
scatterplot_ggrepel(
  data_frame,
  x_index = NULL,
  y_index = NULL,
  label_index = NULL,
  label_color_index = NULL,
  point_shape_index = NULL,
  point_color_index = NULL,
  point_size = 4,
  alpha = 0.5,
  maintitle = NULL,
  xlabel = NULL,
  ylabel = NULL,
  ...
)
}
\arguments{
\item{data_frame}{# df, with at least three columns, 1 as labeling, 2 as x y mapping}

\item{...}{}
}
\value{
a ggplot2 object
}
\description{
scatterplot_ggrepel, a wrapper of scatter plot with ggplot2, and repel function, easy to map color shape etc
}
