% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary_plot_matrix.r
\name{ternary_plot_matrix}
\alias{ternary_plot_matrix}
\title{Biological ternary plot from three rows of the matrix}
\usage{
ternary_plot_matrix(data_matrix, data_meta, three_points)
}
\arguments{
\item{data_matrix}{a two dimentional data matrix, with rows as annation/function/taxon, while columns are samples}

\item{data_meta}{a dataframe with sample-group matching}

\item{three_points}{could be "top3", or a vector of row index, or a vector of row names. If vector, the length must be 3. see examples}
}
\value{
a data.frame of matrix_selected and ggplot2 object of ternary.plot
}
\description{
This is a wrapper of ggern function in ggtern package. The wrapper makes it easier to plot matrix data with grouping information.
See the details for more information, and run the example for a sample plot to have a look at what a ternary plot look like
}
\examples{


}
\seealso{
\code{\link{ggtern}}
}
