% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_protein_table.R
\name{tidy_protein_table}
\alias{tidy_protein_table}
\title{Tidy protein(proteinGroups.txt) or similiar prtoein table}
\usage{
tidy_protein_table(protein_table)
}
\arguments{
\item{protein_table}{}
}
\value{

}
\description{
Tidy protein(proteinGroups.txt) or similiar prtoein table
}
