% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_taxon_table.R
\name{tidy_taxon_table}
\alias{tidy_taxon_table}
\title{tidy taxon table to facilitate downstream analysis}
\usage{
tidy_taxon_table(data_table)
}
\arguments{
\item{data_table}{}
}
\value{
list of values/tables/list
}
\description{
tidy taxon table to facilitate downstream analysis
}
\examples{


}
