% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_upload_table.R
\name{UPLOAD_TABLE_SERVER2}
\alias{UPLOAD_TABLE_SERVER2}
\title{shiny module for table upload, server end, version 2}
\usage{
UPLOAD_TABLE_SERVER2(input, output, session, display_after = TRUE)
}
\arguments{
\item{display_after}{}
}
\value{

}
\description{
shiny module for table upload, server end, version 2
}
