% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_upload_table.R
\name{UPLOAD_TABLE_UI}
\alias{UPLOAD_TABLE_UI}
\title{shiny module for table upload, ui end}
\usage{
UPLOAD_TABLE_UI(
  id,
  header = TRUE,
  rownames = TRUE,
  boxwidth = 12,
  boxtitle = "Upload data file (tsv/csv)"
)
}
\arguments{
\item{boxtitle}{}
}
\value{

}
\description{
UI includes table upload and display, with box and box with 4-8 ratio
}
