% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_upload_table.R
\name{UPLOAD_TABLE_UI2}
\alias{UPLOAD_TABLE_UI2}
\title{shiny module for table upload, ui end, version2}
\usage{
UPLOAD_TABLE_UI2(
  id,
  header = TRUE,
  rownames = TRUE,
  boxwidth = 12,
  check.names = TRUE,
  collapsed = FALSE,
  boxtitle = "Upload data file (tsv/csv)"
)
}
\arguments{
\item{boxtitle}{}
}
\value{

}
\description{
ui2 is a slight modification of the version 1, with addition of check.names option, while removing the quote options
}
