% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withConsoleRedirect.R
\name{withConsoleRedirect}
\alias{withConsoleRedirect}
\title{redirect R output to the browser}
\usage{
withConsoleRedirect(containerId, expr, type = "message")
}
\arguments{
\item{containerId}{shiny container id}

\item{type}{Default is message, could be "output", see type in capture.output, passed to sink(), ?sink for more information}
}
\value{

}
\description{
redirect R output to the browser
}
\examples{

# NOT run:
# UI end:
# verbatimTextOutput("console_error1",placeholder = FALSE),
# verbatimTextOutput("console_out1", placeholder = FALSE
#
# SERVER END:
# observeEvent(input$Apply1, {
#   withConsoleRedirect("console_error1", try(log("a")))
# })
#
# observeEvent(input$Apply2, {
#   withConsoleRedirect("console_out1",cat("Result of log(10):"), type = "output") # if you want to catch up the output, set type to ouput
# })

}
