% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_lipid_combo.R
\name{find_lipid_combo}
\alias{find_lipid_combo}
\title{find 3 chains combo for a lipid structure}
\usage{
find_lipid_combo(combo_target, chains_observed)
}
\arguments{
\item{combo_target}{a string, general format as carbon number:double bound number, like 51:4}

\item{chains_observed}{a string vector, consists of all observed fatty acid chains, with  general format as carbon number:double bound number,}
}
\value{
a data.frame, with all possible 3 chain compositions. The order of the chain does not matter.
}
\description{
find_lipid_combo start with more than 3 observed fatty acid chains with defined carbo and double bound number to find the
combo to achieve a target lipid structure. The three chains could be the same (replacement sampling).
}
\examples{

combo_target <- "51:4"
chains_observed <- c("18:1","18:2","18:3", "17:1", "17:2" ,"16:0" ,"16:1", "15:0" ,"15:1")

r <- find_lipid_combo(combo_target,chains_observed)

#error run
r <- find_lipid_combo(combo_target,"dat")
r <- find_lipid_combo(combo_target,c("18:1", "bbb", "17:1", "17:2") ) #notworking, can run, but result is not right
}
\seealso{
\code{\link{combn}}
}
