% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_plot_3d_interactive_3.R
\name{PCA_plot_3d_interactive_3}
\alias{PCA_plot_3d_interactive_3}
\title{Title}
\usage{
PCA_plot_3d_interactive_3(prcomp_out, grouping)
}
\arguments{
\item{grouping}{}
}
\value{

}
\description{
Title
}
