% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_plot_with_confidence_2.R
\name{PCA_plot_with_confidence_2}
\alias{PCA_plot_with_confidence_2}
\title{internal method for displaying pca result, with confidence eclipse}
\usage{
PCA_plot_with_confidence_2(prcomp_out, data_meta)
}
\arguments{
\item{prcomp_out}{output from prcomp}

\item{data_meta}{grouping information to label with groups, format is required as data.frame, with first column as sample name, second column as group}
}
\value{

}
\description{
internal method for displaying pca result, with confidence eclipse
}
