% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_plot_with_ellipse_kmeans.R
\name{PCA_plot_with_ellipse_kmeans_2}
\alias{PCA_plot_with_ellipse_kmeans_2}
\title{Title}
\usage{
PCA_plot_with_ellipse_kmeans_2(prcomp_out, grouping)
}
\arguments{
\item{grouping}{}
}
\value{

}
\description{
Title
}
