% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_wrapper_prcomp2.R
\name{PCA_wrapper_prcomp2}
\alias{PCA_wrapper_prcomp2}
\title{wrapper function for PCA with more options to preprocess data and output plots}
\usage{
PCA_wrapper_prcomp2(data_matrix, data_meta, inputation = FALSE, Q = 0.75)
}
\arguments{
\item{data_matrix}{target data matrix for PCA analysis, a long table, with rows as features, while columns as observation/experiments}

\item{data_meta}{a data.frame with two columns, first as experiment, second as grouping}

\item{inputation}{logic, if do imputation. default as FALSE,}

\item{Q}{a value in [0, 1], if do filtering before analysis. default as 0.75, meaning only keep rows with more than 75% present values. Too many missing value will distart PCA analysis}
}
\value{
a list,including direct PCA result and plot
}
\description{
This wrapper uses prcomp as the method to do pca. This is preferred for omic data, with more features(lines) than samples(columns)
}
\examples{

test_data <- generate_test_data()
p <- PCA_wrapper_prcomp2(data_matrix = test_data$matrix, data_meta = test_data$meta, inputation = TRUE, Q = 0.75)

}
