% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary_plot_matrix.r
\name{ternary_plot_matrix}
\alias{ternary_plot_matrix}
\title{Biological ternary plot from three rows of the matrix}
\usage{
ternary_plot_matrix(data_matrix, data_meta, three_points)
}
\arguments{
\item{data_matrix}{a two dimentional data matrix, with rows as annation/function/taxon, while columns are samples}

\item{data_meta}{a dataframe with sample-group matching}

\item{three_points}{could be "top3", or a vector of row index, or a vector of row names. If vector, the length must be 3. see examples}
}
\value{
a data.frame of matrix_selected and ggplot2 object of ternary.plot
}
\description{
This is a wrapper of ggern function in ggtern package. The wrapper makes it easier to plot matrix data with grouping information.
See the details for more information, and run the example for a sample plot to have a look at what a ternary plot look like
}
\examples{
#data preparation

 matrix_test <- matrix(runif(2000), nrow = 20)
 rownames(matrix_test) <- paste("COG", 1:20, sep = "_")
 colnames(matrix_test) <- paste("sample", 1:100, sep = "_")
 factor_test <- sample(LETTERS[1:3], 100, replace = TRUE)
 meta_test <- data.frame(samplename = colnames(matrix_test), grouping = factor_test)

#Plot in three ways

t <- ternary_plot_matrix(data_matrix = matrix_test, data_meta = meta_test, three_points = c("COG_1", "COG_2", "COG_3"))
t$ternary.plot
t <- ternary_plot_matrix(data_matrix = matrix_test, data_meta = meta_test, three_points = "top3")
t$ternary.plot
t <- ternary_plot_matrix(data_matrix = matrix_test, data_meta = meta_test, three_points = c(2,4,5))
t$ternary.plot


}
\seealso{
\code{\link{ggtern}}
}
