% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_peptide_table.R
\name{tidy_peptide_table}
\alias{tidy_peptide_table}
\title{clean and sort the peptide table input into a list.}
\usage{
tidy_peptide_table(peptide.txt)
}
\arguments{
\item{peptide.txt}{a data.frame read from peptide.txt}
}
\value{
an object list, including a peptide intensity matrix, vectors of peptide_sequence,length, misscleavage
}
\description{
Cleaning include removing contaminant, reverse hits, extrat the peptide intensity matrix ready to use,
extract the experiment design, put the peptide length, sequence, missing cleavage site into vectors ready to use
}
\examples{
 file_path <- system.file("extdata","peptides_1000.txt", package = "metalab")
 peptide_table <- read.delim(file_path, header = TRUE,check.names = FALSE, stringsAsFactors = FALSE) #
 t <- tidy_peptide_table(peptide_table)
 t$intensity_matrix
 t$summary_rawfiles
 t$experiment

}
